feat(json): point数据coord 设置支持数组
