+++ src/component/control/index.js
@@ -1,11 +1,11 @@
-import Control from './base';
+import Base from './base';
 import Zoom from './zoom';
 import Scale from './Scale';
 import Attribution from './attribution';
 import Layers from './layer';
 
 export {
-  Control,
+  Base,
   Zoom,
   Scale,
   Attribution,
